Fix broken WebWorker URL when on GitHub pages
Fixes an issue where the chart function was failing to resolve its data
frame WebWorker due to relative path issues when the site is deployed
with a context path.

The fix is specifically for GitHub pages, which assumes the site is
deployed at the root of the mandatory repository context path.